fix: 버튼 컴포넌트 수정 및 템플릿 수정
