+++ src/components/Button/Button/Button.module.scss
@@ -136,16 +136,3 @@
   align-items: center;
   gap: 0.75rem;
 }
-
-.dashboardText {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.dashboardName {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 0.75rem;
-}
