--- src/components/Button/ButtonAddCard/ButtonAddCard.module.scss
@@ -1,19 +1,0 @@
-.buttonAddCard {
-  cursor: pointer;
-  border: none;
-  background: $color-white-ffffff;
-  width: 19.75rem;
-  height: 2.5rem;
-  border-radius: 0.5rem;
-  border: 0.0625rem solid $color-gray-d9d9d9;
-  color: $color-black-333236;
-
-  @include tablet {
-    width: 34rem;
-  }
-
-  @include mobile {
-    width: 17.75rem;
-    height: 2rem;
-  }
-}
