--- src/components/Button/ButtonAddColumn/ButtonAddColumn.module.scss
@@ -1,22 +1,0 @@
-.buttonAddColumn {
-  cursor: pointer;
-  border: none;
-  background: $color-white-ffffff;
-  width: 22.25rem;
-  height: 4.5rem;
-  border-radius: 0.5rem;
-  border: 0.0625rem solid $color-gray-d9d9d9;
-  font-size: 1.25rem;
-  font-weight: 700;
-  color: $color-black-333236;
-
-  @include tablet {
-    width: 34rem;
-  }
-
-  @include mobile {
-    width: 17.75rem;
-    height: 4rem;
-    font-size: 1rem;
-  }
-}
