+++ src/components/Button/ButtonDashboard/ButtonDashboard.module.scss
@@ -1,6 +1,4 @@
-.buttonDashboard {
-  cursor: pointer;
-  border: none;
+.dashboard {
   background: $color-gray-fafafa;
   width: 20rem;
   height: 4rem;
@@ -15,4 +13,23 @@
     height: 3.25rem;
     font-size: 1rem;
   }
+}
+
+.dashboardText {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.dashboardName {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.circle {
+  border-radius: 50%;
+  width: 0.5rem;
+  height: 0.5rem;
 }
