--- src/components/Button/ButtonDelete/ButtonDelete.module.scss
@@ -1,18 +1,0 @@
-.buttonDelete {
-  cursor: pointer;
-  border: none;
-  background: $color-white-ffffff;
-  width: 5.25rem;
-  height: 2rem;
-  border-radius: 0.25rem;
-  border: 0.0625rem solid $color-gray-d9d9d9;
-  font-size: 1rem;
-  font-weight: 500;
-  color: $color-violet-5534da;
-
-  @include mobile {
-    width: 3.25rem;
-    height: 1.75rem;
-    font-size: 0.75rem;
-  }
-}
